ct stuff

+++ src/preproc/anatomy.py
@@ -995,7 +995,7 @@ def create_labels_names_lookup(subject, atlas):
     return lookup
 
 
-def create_new_subject_blend_file(subject, atlas, overwrite_blend=False):
+def create_new_subject_blend_file(subject, atlas, overwrite_blend=False, ask_if_overwrite_blend=True):
     # Create a file for the new subject
     atlas = utils.get_real_atlas_name(atlas, short_name=True)
     new_fname = op.join(MMVT_DIR, '{}_{}.blend'.format(subject, atlas))
@@ -1003,7 +1003,7 @@ def create_new_subject_blend_file(subject, atlas, overwrite_blend=False):
     if not op.isfile(empty_subject_fname):
         resources_dir = op.join(utils.get_parent_fol(levels=2), 'resources')
         shutil.copy(op.join(resources_dir, 'empty_subject.blend'), empty_subject_fname)
-    if op.isfile(new_fname) and not overwrite_blend:
+    if op.isfile(new_fname) and not overwrite_blend and args.ask_if_overwrite_blend:
         overwrite = input('The file {} already exist, do you want to overwrite? '.format(new_fname))
         if au.is_true(overwrite):
            os.remove(new_fname)
@@ -1342,7 +1342,7 @@ def main(subject, remote_subject_dir, args, flags):
 
     if utils.should_run(args, 'create_new_subject_blend_file'):
         flags['create_new_subject_blend_file'] = create_new_subject_blend_file(
-            subject, args.atlas, args.overwrite_blend)
+            subject, args.atlas, args.overwrite_blend, args.ask_if_overwrite_blend)
 
     if 'cerebellum_segmentation' in args.function:
         flags['save_cerebellum_coloring'] = save_cerebellum_coloring(subject)
@@ -1395,6 +1395,7 @@ def read_cmd_args(argv=None):
     parser.add_argument('--overwrite_faces_verts', help='overwrite_faces_verts', required=False, default=0, type=au.is_true)
     parser.add_argument('--overwrite_ply_files', help='overwrite_ply_files', required=False, default=0, type=au.is_true)
     parser.add_argument('--overwrite_blend', help='overwrite_blend', required=False, default=0, type=au.is_true)
+    parser.add_argument('--ask_if_overwrite_blend', help='', required=False, default=1, type=au.is_true)
     parser.add_argument('--overwrite_flat_surf', help='overwrite_flat_surf', required=False, default=0, type=au.is_true)
     parser.add_argument('--solve_labels_collisions', help='solve_labels_collisions', required=False, default=0, type=au.is_true)
     parser.add_argument('--morph_labels_from_fsaverage', help='morph_labels_from_fsaverage', required=False, default=1, type=au.is_true)
